output/services/registry-service/sdk/pyproject.toml
[project]
name = "venturestrat-registry"
version = "0.1.0"
description = "Python SDK for VentureStrat Registry Service"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "VentureStrat Team"}
]
keywords = ["venturestrat", "registry", "service-discovery", "microservices"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "httpx>=0.26.0",
    "grpcio>=1.60.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["venturestrat_registry"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
